List of changes
Major changes: support for "Point mass" added to the interface
and tested. The point mass is placed under the geometry root.
The scope for the MPC attachment is a face or a group of faces.
Support tested, including persistance.

Minor changes: code simplifications where possible (old pieces
of program, not using helpers)

To do:
- support for input file generation and input file test.
- support for the remote attachment to edge/edges or point/points